requirements.txt
backcall==0.1.0
colorama==0.4.1
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-extensions==2.2.5
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.4
Pygments==2.4.2
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
whitenoise==4.1.4
